+++ methods/src/main/kotlin/methods/polynomials/PolynomialsPlans.kt
@@ -4,6 +4,7 @@ import engine.context.ResourceData
 import engine.expressions.Minus
 import engine.expressions.Power
 import engine.expressions.Product
+import engine.expressions.ValueExpression
 import engine.methods.CompositeMethod
 import engine.methods.PublicMethod
 import engine.methods.RunnerMethod
@@ -50,6 +51,7 @@ enum class PolynomialsPlans(override val runner: CompositeMethod) : RunnerMethod
         plan {
             explanation = Explanation.SimplifyPolynomialExpressionInOneVariable
             specificPlans(ConstantExpressionsPlans.SimplifyConstantExpression)
+            pattern = condition { it is ValueExpression }
 
             steps {
                 whilePossible { deeply(simpleTidyUpSteps) }